add an updated version of tumor UMI proportion in the mixture model; move E step related functions in utils_hmm
